@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RuntimeIdentifiers>linux-x64;linux-arm64;osx-x64;osx-arm64;win-x64;win-arm64</RuntimeIdentifiers>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -12,17 +12,28 @@
     <Copyright>Copyright © Alex Thomson</Copyright>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(CI)' == 'true' or '$(DOTNET_RUNNING_IN_CONTAINER)' == 'true'">
+    <JsonSerializerIsReflectionEnabledByDefault>true</JsonSerializerIsReflectionEnabledByDefault>
+    <PublishSelfContained>true</PublishSelfContained>
+    <PublishSingleFile>true</PublishSingleFile>
+    <PublishTrimmed>true</PublishTrimmed>
+    <SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings>
+    <TrimMode>partial</TrimMode>
+  </PropertyGroup>
+
   <ItemGroup>
-    <PackageReference Include="Discord.Net.Webhook" Version="3.10.0" />
+    <PackageReference Include="Discord.Net.Webhook" Version="3.13.0" />
     <PackageReference Include="Docker.DotNet" Version="3.125.15" />
     <PackageReference Include="Humanizer.Core" Version="2.14.1" />
-    <PackageReference Include="LXGaming.Common.Hosting" Version="1.5.1" />
-    <PackageReference Include="LXGaming.Common.Serilog" Version="1.5.1" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
-    <PackageReference Include="Serilog.Extensions.Hosting" Version="7.0.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
+    <PackageReference Include="LXGaming.Common" Version="1.8.1" />
+    <PackageReference Include="LXGaming.Common.Hosting" Version="1.8.1" />
+    <PackageReference Include="LXGaming.Common.Serilog" Version="1.8.1" />
+    <PackageReference Include="LXGaming.Configuration" Version="1.1.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
-    <PackageReference Include="Serilog.Sinks.File.Archive" Version="1.0.3" />
+    <PackageReference Include="Serilog.Sinks.File.Archive" Version="1.0.4" />
   </ItemGroup>
 
 </Project>
