@@ -1,8 +1,11 @@
 ﻿using System.IO.Compression;
 using System.Reflection;
+using System.Text.Json;
 using LXGaming.Captain.Configuration;
 using LXGaming.Common.Hosting;
 using LXGaming.Common.Serilog;
+using LXGaming.Configuration;
+using LXGaming.Configuration.Json;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Serilog;
@@ -21,11 +24,15 @@ Log.Logger = new LoggerConfiguration()
         hooks: new ArchiveHooks(CompressionLevel.Optimal))
     .CreateBootstrapLogger();
 
-Log.Information("Initializing...");
+Log.Information("Initialising...");
 
 try {
-    var configuration = new JsonConfiguration(Directory.GetCurrentDirectory());
-    await configuration.LoadConfigurationAsync();
+    var configuration = new DefaultConfiguration();
+    await configuration.LoadJsonFileAsync<Config>(
+        options: new JsonSerializerOptions {
+            WriteIndented = true
+        }
+    );
 
     var builder = Host.CreateDefaultBuilder(args);
     builder.UseSerilog();
@@ -40,7 +47,7 @@ try {
     await host.RunAsync();
     return 0;
 } catch (Exception ex) {
-    Log.Fatal(ex, "Application failed to initialize");
+    Log.Fatal(ex, "Application failed to initialise");
     return 1;
 } finally {
     Log.CloseAndFlush();
