@@ -6,22 +6,14 @@ using Microsoft.Extensions.Logging;
 namespace LXGaming.Captain.Services.Notification;
 
 [Service(ServiceLifetime.Singleton)]
-public class NotificationService {
-
-    private readonly ILogger<NotificationService> _logger;
-    private readonly IServiceProvider _serviceProvider;
-
-    public NotificationService(ILogger<NotificationService> logger, IServiceProvider serviceProvider) {
-        _logger = logger;
-        _serviceProvider = serviceProvider;
-    }
+public class NotificationService(ILogger<NotificationService> logger, IServiceProvider serviceProvider) {
 
     public async Task NotifyAsync(Func<INotificationProvider, Task> func) {
-        foreach (var notification in _serviceProvider.GetServices<INotificationProvider>()) {
+        foreach (var notification in serviceProvider.GetServices<INotificationProvider>()) {
             try {
                 await func(notification);
             } catch (Exception ex) {
-                _logger.LogError(ex, "Encountered an error while sending notification");
+                logger.LogError(ex, "Encountered an error while notifying {Name}", notification.GetType().Name);
             }
         }
     }
