@@ -4,6 +4,7 @@ using LXGaming.Captain.Configuration;
 using LXGaming.Captain.Services.Docker.Models;
 using LXGaming.Captain.Utilities;
 using LXGaming.Common.Hosting;
+using LXGaming.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -11,30 +12,26 @@ using Microsoft.Extensions.Logging;
 namespace LXGaming.Captain.Services.Notification.Providers;
 
 [Service(ServiceLifetime.Singleton, typeof(INotificationProvider))]
-public class DiscordNotificationProvider : IHostedService, INotificationProvider {
+public class DiscordNotificationProvider(
+    IConfiguration configuration,
+    ILogger<DiscordNotificationProvider> logger) : IHostedService, INotificationProvider {
 
-    private readonly IConfiguration _configuration;
-    private readonly ILogger<DiscordNotificationProvider> _logger;
+    private readonly IProvider<Config> _config = configuration.GetRequiredProvider<Config>();
     private DiscordWebhookClient? _discordClient;
 
-    public DiscordNotificationProvider(IConfiguration configuration, ILogger<DiscordNotificationProvider> logger) {
-        _configuration = configuration;
-        _logger = logger;
-    }
-
     public Task StartAsync(CancellationToken cancellationToken) {
-        var discordCategory = _configuration.Config?.NotificationCategory;
-        if (discordCategory == null) {
+        var category = _config.Value?.NotificationCategory;
+        if (category == null) {
             throw new InvalidOperationException("NotificationCategory is unavailable");
         }
 
-        if (!discordCategory.Enabled) {
+        if (!category.Enabled) {
             return Task.CompletedTask;
         }
 
-        var url = discordCategory.Url;
+        var url = category.Url;
         if (string.IsNullOrEmpty(url)) {
-            _logger.LogWarning("Url has not been configured for Discord");
+            logger.LogWarning("Url has not been configured for Discord");
             return Task.CompletedTask;
         }
 
@@ -47,21 +44,11 @@ public class DiscordNotificationProvider : IHostedService, INotificationProvider
         return Task.CompletedTask;
     }
 
-    public Task SendAlertAsync(Embed embed) {
-        var mentions = _configuration.Config?.NotificationCategory.Mentions;
-        return SendMessageAsync(mentions != null ? string.Join(' ', mentions) : null, embed);
-    }
-
-    public Task SendMessageAsync(string? text = null, Embed? embed = null) {
-        var discordCategory = _configuration.Config?.NotificationCategory;
-        return _discordClient?.SendMessageAsync(
-            text: text,
-            embeds: new[] { embed },
-            username: discordCategory?.Username,
-            avatarUrl: discordCategory?.AvatarUrl) ?? Task.CompletedTask;
-    }
-
     public Task SendHealthStatusAsync(Container container, bool state) {
+        if (_discordClient == null) {
+            return Task.CompletedTask;
+        }
+
         var embedBuilder = new EmbedBuilder();
         embedBuilder.WithColor(state ? Color.Green : Color.Red);
         embedBuilder.WithTitle("Health");
@@ -69,10 +56,14 @@ public class DiscordNotificationProvider : IHostedService, INotificationProvider
         embedBuilder.AddField("Name", $"```\n{container.Name}\n```", true);
         embedBuilder.AddField("Status", $"```\n{(state ? "Healthy" : "Unhealthy")}\n```", true);
         embedBuilder.WithFooter($"{Constants.Application.Name} v{Constants.Application.Version}");
-        return SendAlertAsync(embedBuilder.Build());
+        return SendAlertAsync(new[] { embedBuilder.Build() });
     }
 
     public Task SendLogAsync(Container container, string message) {
+        if (_discordClient == null) {
+            return Task.CompletedTask;
+        }
+
         var embedBuilder = new EmbedBuilder();
         embedBuilder.WithColor(Color.Orange);
         embedBuilder.WithTitle("Log");
@@ -80,10 +71,14 @@ public class DiscordNotificationProvider : IHostedService, INotificationProvider
         embedBuilder.AddField("Name", $"```\n{container.Name}\n```", true);
         embedBuilder.AddField("Message", $"```\n{message}\n```", true);
         embedBuilder.WithFooter($"{Constants.Application.Name} v{Constants.Application.Version}");
-        return SendAlertAsync(embedBuilder.Build());
+        return SendAlertAsync(new[] { embedBuilder.Build() });
     }
 
     public Task SendRestartLoopAsync(Container container, string exitCode) {
+        if (_discordClient == null) {
+            return Task.CompletedTask;
+        }
+
         var embedBuilder = new EmbedBuilder();
         embedBuilder.WithColor(Color.Orange);
         embedBuilder.WithTitle("Restart Loop Detected");
@@ -91,6 +86,33 @@ public class DiscordNotificationProvider : IHostedService, INotificationProvider
         embedBuilder.AddField("Name", $"```\n{container.Name}\n```", true);
         embedBuilder.AddField("Exit Code", $"```\n{exitCode}\n```", true);
         embedBuilder.WithFooter($"{Constants.Application.Name} v{Constants.Application.Version}");
-        return SendAlertAsync(embedBuilder.Build());
+        return SendAlertAsync(new[] { embedBuilder.Build() });
+    }
+
+    private Task<ulong> SendAlertAsync(IEnumerable<Embed>? embeds = null) {
+        var mentions = _config.Value?.NotificationCategory.Mentions;
+        if (mentions == null || mentions.Count == 0) {
+            return SendMessageAsync(null, embeds);
+        }
+
+        return SendMessageAsync(string.Join(' ', mentions), embeds);
+    }
+
+    private Task<ulong> SendMessageAsync(string? text = null, IEnumerable<Embed>? embeds = null) {
+        var category = _config.Value?.NotificationCategory;
+        if (category == null) {
+            throw new InvalidOperationException("NotificationCategory is unavailable");
+        }
+
+        if (_discordClient == null) {
+            throw new InvalidOperationException("DiscordClient is unavailable");
+        }
+
+        return _discordClient.SendMessageAsync(
+            text: text,
+            embeds: embeds,
+            username: category.Username,
+            avatarUrl: category.AvatarUrl
+        );
     }
 }
