@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using LXGaming.Common.Utilities;
 
 namespace LXGaming.Captain.Utilities;
 
@@ -10,7 +11,7 @@ public static class Constants {
         public const string Authors = "LX_Gaming";
         public const string Website = "https://lxgaming.github.io/";
 
-        public static readonly string Version = Toolbox.GetAssemblyVersion(Assembly.GetExecutingAssembly());
+        public static readonly string Version = AssemblyUtils.GetAssemblyVersion(Assembly.GetExecutingAssembly());
         public static readonly string UserAgent = Name + "/" + Version + " (+" + Website + ")";
     }
 }
